fix: clean up meet entity and related dtos
delete test controller, remove unused getter, replace ownerId,
rename meetDt to meetAt, update integer to long type,
implement MeetSummaryDTO
