+++ src/main/java/org/ktb/modie/presentation/v1/controller/AuthController.java
@@ -1,0 +1,4 @@
+package org.ktb.modie.presentation.v1.controller;
+
+public class AuthController {
+}
