+++ src/main/java/org/ktb/modie/presentation/v1/dto/UserDto.java
@@ -1,0 +1,5 @@
+package org.ktb.modie.presentation.v1.dto;
+
+public class UserDto {
+
+}
