--- src/main/java/org/ktb/modie/v1/controller/AuthController.java
@@ -1,4 +1,0 @@
-package org.ktb.modie.v1.controller;
-
-public class AuthController {
-}
