--- src/main/java/org/ktb/modie/v1/dto/UserDto.java
@@ -1,5 +1,0 @@
-package org.ktb.modie.v1.dto;
-
-public class UserDto {
-
-}
